+++ components/content/write-content/content-form/content-form.tsx
@@ -5,7 +5,7 @@ import Button from '@components/generic/button/button'
 import Card from '@components/generic/card/card'
 import ContentSummaryView from '../../read-content/content-summary-view/content-summary-view'
 import DynamicField from '@components/generic/dynamic-field/dynamic-field-edit'
-import { FIELDS } from '@lib/config/config-constants'
+import { FIELDS } from '@lib/constants'
 import GroupSummaryView from '@components/groups/read/group-summary-view/group-summary-view'
 import Link from 'next/link'
 import Toggle from '@components/generic/toggle/toggle'
