+++ components/groups/write/group-form/group-form.js
@@ -5,7 +5,7 @@ import Button from '@components/generic/button/button'
 import DynamicField from '@components/generic/dynamic-field/dynamic-field-edit'
 import Toggle from '@components/generic/toggle/toggle'
 import fetch from '@lib/fetcher'
-import { FIELDS } from '@lib/config/config-constants'
+import { FIELDS } from '@lib/constants'
 import GroupSummaryView from '../../read/group-summary-view/group-summary-view'
 import Link from 'next/link'
 import UserPermissions from './user-permissions/user-permissions'
