+++ lib/api/api-helpers/dynamic-file-upload.js
@@ -1,4 +1,4 @@
-import { FIELDS } from '@lib/config/config-constants'
+import { FIELDS } from '@lib/constants'
 import { deleteFile } from '@lib/api/storage'
 import merge from 'deepmerge'
 import { uploadFile } from '../storage'
