+++ lib/api/db/index.js
@@ -1,4 +1,4 @@
-import { DATABASES } from '../../config/config-constants'
+import { DATABASES } from '@lib/constants'
 import config from '../../config'
 import { connect as connectFirebase } from './firestoreDB'
 import { connect as connectInMemoryDB } from './inMemoryDB'
