+++ lib/api/storage/index.js
@@ -1,7 +1,7 @@
 import AWSStorage from './aws-storage'
 import FirebaseStorage from './firestore-storage'
 import GoogleStorage from './google-storage'
-import { STORAGES } from '../../config/config-constants'
+import { STORAGES } from '@lib/constants'
 import config from '../../config'
 
 let storage = () => console.error('Storage is not implemented')
