+++ lib/config/config.schema.js
@@ -1,6 +1,6 @@
 const { object, string, number, mixed, array, boolean } = require('yup')
 
-import { DATABASES, FIELDS } from './config-constants'
+import { DATABASES, FIELDS } from '@lib/constants'
 
 import { UserSchema } from '@lib/api/entities/users/user.schema'
