+++ lib/config/load-config.js
@@ -1,8 +1,17 @@
-import { DATABASES, ROLES } from './config-constants'
-
+import { DATABASES } from '@lib/constants'
 import { getConfig } from '../../edge.config'
 import merge from 'deepmerge'
 import schema from './config.schema'
+
+const ROLES = {
+  ADMIN: 'ADMIN',
+  PUBLIC: 'PUBLIC',
+  USER: 'USER'
+}
+
+
+
+
 
 const defaultConfig = {
   title: '', // Empty to throw error if not configured
