+++ lib/constants/fields.ts
@@ -1,20 +1,3 @@
-export const DATABASES = {
-  IN_MEMORY: 'IN_MEMORY',
-  FIREBASE: 'FIREBASE',
-  MONGO: 'MONGO',
-}
-
-export const ROLES = {
-  ADMIN: 'ADMIN',
-  USER: 'USER',
-  PUBLIC: 'PUBLIC',
-}
-
-export const STORAGES = {
-  AWS: 'AWS',
-  FIREBASE: 'FIREBASE',
-  GOOGLE: 'GOOGLE',
-}
 
 export const FIELDS = {
   TEXT: 'text',
