+++ lib/constants/storages.ts
@@ -1,0 +1,6 @@
+
+export const STORAGES = {
+  AWS: 'AWS',
+  FIREBASE: 'FIREBASE',
+  GOOGLE: 'GOOGLE',
+}
