+++ lib/validations/fields.js
@@ -1,4 +1,4 @@
-import { FIELDS } from '../config/config-constants'
+import { FIELDS } from '@lib/constants'
 
 export const validateField = (name, value, fieldDefinition) => {
   const errors = []
